move getNewFileMock to test_utils and remove unused param
