+++ apps/test/unit/codebridge/FileBrowser/prompts/openNewFilePromptTest.ts
@@ -1,24 +1,14 @@
-import {NewFileFunction} from '@codebridge/codebridgeContext/types';
 import {DEFAULT_FOLDER_ID} from '@codebridge/constants';
 import {openNewFilePrompt} from '@codebridge/FileBrowser/prompts/openNewFilePrompt';
-import {FolderId, ProjectFile} from '@codebridge/types';
 
 import {EVENTS} from '@cdo/apps/metrics/AnalyticsConstants';
 
 import {testProject, validationFile} from '../../test-files/';
-import {getDialogControlMock, getAnalyticsMock} from '../../test_utils';
-
-const getNewFileMock = (parentId: FolderId): [ProjectFile, NewFileFunction] => {
-  const newFileData = {} as ProjectFile;
-  const mock: NewFileFunction = ({fileName, folderId}) => {
-    newFileData.name = fileName;
-    if (folderId) {
-      newFileData.folderId = folderId;
-    }
-  };
-
-  return [newFileData, mock];
-};
+import {
+  getDialogControlMock,
+  getAnalyticsMock,
+  getNewFileMock,
+} from '../../test_utils';
 
 describe('openNewFilePrompt', function () {
   it('can successfully add a new file to root w/o validation file', async function () {
@@ -26,7 +16,7 @@ describe('openNewFilePrompt', function () {
     const newFileName = 'valid_file.txt';
     const folderId = DEFAULT_FOLDER_ID;
 
-    const [newFileData, newFileDataMock] = getNewFileMock(folderId);
+    const [newFileData, newFileDataMock] = getNewFileMock();
 
     await openNewFilePrompt({
       folderId,
@@ -49,7 +39,7 @@ describe('openNewFilePrompt', function () {
     const newFileName = 'valid_file.txt';
     const folderId = DEFAULT_FOLDER_ID;
 
-    const [newFileData, newFileDataMock] = getNewFileMock(folderId);
+    const [newFileData, newFileDataMock] = getNewFileMock();
 
     await openNewFilePrompt({
       dialogControl: getDialogControlMock(newFileName),
@@ -71,7 +61,7 @@ describe('openNewFilePrompt', function () {
     const newFileName = 'valid_file.txt';
     const folderId = '1';
 
-    const [newFileData, newFileDataMock] = getNewFileMock(folderId);
+    const [newFileData, newFileDataMock] = getNewFileMock();
 
     await openNewFilePrompt({
       folderId,
@@ -94,7 +84,7 @@ describe('openNewFilePrompt', function () {
     const newFileName = 'valid_file.txt';
     const folderId = DEFAULT_FOLDER_ID;
 
-    const [newFileData, newFileDataMock] = getNewFileMock(folderId);
+    const [newFileData, newFileDataMock] = getNewFileMock();
 
     await openNewFilePrompt({
       folderId,
@@ -117,7 +107,7 @@ describe('openNewFilePrompt', function () {
     const newFileName = 'invalid_file';
     const folderId = DEFAULT_FOLDER_ID;
 
-    const [newFileData, newFileDataMock] = getNewFileMock(folderId);
+    const [newFileData, newFileDataMock] = getNewFileMock();
 
     await openNewFilePrompt({
       folderId,
